+++ lib/ckb_addrss.dart
@@ -1,3 +1,5 @@
 library ckb_address;
 
 export 'src/address/address_generator.dart';
+export 'src/address/address_parser.dart';
+export 'src/address/address_type.dart';