+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.30.3
+
+### Reactor
+
+- Add fromBytes to UInt32
+- Rename cell collector
+
+## 0.30.2
+
+### Feature
+
+- Adjust core and address export
+
 ## 0.30.1
 
 ### Feature